+++ tests/table_spec.lua
@@ -10,34 +10,35 @@ describe('table.md', function()
 
         vim.list_extend(expected, {
             util.heading(row:get(), 1),
-            util.table_border(row:increment(2), 'above', { 11, 24 }),
+            util.table_border(row:increment(2), true, { 11, 24 }),
             util.table_pipe(row:get(), 0, true),
             util.table_pipe(row:get(), 12, true),
+            util.table_padding(row:get(), 14, 13),
             util.inline_code(row:get(), 14, 25),
-            util.table_padding(row:get(), 37, 2),
+            util.conceal(row:get(), 26, 37),
             util.table_pipe(row:get(), 37, true),
-            util.table_border(row:increment(), 'delimiter', { 11, 24 }),
+            util.table_delimiter(row:increment(), { 11, { 23, 1 } }),
             util.table_pipe(row:increment(), 0, false),
             util.inline_code(row:get(), 2, 12),
             util.table_padding(row:get(), 13, 2),
             util.table_pipe(row:get(), 13, false),
+            util.table_padding(row:get(), 15, 16),
             util.link(row:get(), 15, 38, 'web'),
-            util.table_padding(row:get(), 39, 16),
             util.table_pipe(row:get(), 39, false),
-            util.table_border(row:get(), 'below', { 11, 24 }),
+            util.table_border(row:get(), false, { 11, 24 }),
         })
 
         vim.list_extend(expected, {
             util.heading(row:increment(2), 1),
-            util.table_border(row:increment(2), 'above', { 11, 11 }),
+            util.table_border(row:increment(2), true, { 11, 11 }),
             util.table_pipe(row:get(), 0, true),
             util.table_pipe(row:get(), 12, true),
             util.table_pipe(row:get(), 24, true),
-            util.table_border(row:increment(), 'delimiter', { 11, 11 }),
+            util.table_delimiter(row:increment(), { 11, 11 }),
             util.table_pipe(row:increment(), 0, false),
             util.table_pipe(row:get(), 12, false),
             util.table_pipe(row:get(), 24, false),
-            util.table_border(row:get(), 'below', { 11, 11 }),
+            util.table_border(row:get(), false, { 11, 11 }),
         })
 
         local actual = util.get_actual_marks()
@@ -51,34 +52,34 @@ describe('table.md', function()
 
         vim.list_extend(expected, {
             util.heading(row:get(), 1),
-            util.table_border(row:increment(2), 'above', { 11, 11 }),
+            util.table_border(row:increment(2), true, { 11, 11 }),
             util.table_pipe(row:get(), 0, true),
             util.table_pipe(row:get(), 12, true),
             util.inline_code(row:get(), 14, 25),
             util.conceal(row:get(), 26, 37),
             util.table_pipe(row:get(), 37, true),
-            util.table_border(row:increment(), 'delimiter', { 11, 11 }, nil, 13),
+            util.table_delimiter(row:increment(), { 11, { 10, 1 } }, 13),
             util.table_pipe(row:increment(), 0, false),
             util.inline_code(row:get(), 2, 12),
             util.table_padding(row:get(), 13, 2),
             util.table_pipe(row:get(), 13, false),
+            util.table_padding(row:get(), 15, 3),
             util.link(row:get(), 15, 38, 'web'),
-            util.table_padding(row:get(), 39, 3),
             util.table_pipe(row:get(), 39, false),
-            util.table_border(row:get(), 'below', { 11, 11 }),
+            util.table_border(row:get(), false, { 11, 11 }),
         })
 
         vim.list_extend(expected, {
             util.heading(row:increment(2), 1),
-            util.table_border(row:increment(2), 'above', { 11, 11 }),
+            util.table_border(row:increment(2), true, { 11, 11 }),
             util.table_pipe(row:get(), 0, true),
             util.table_pipe(row:get(), 12, true),
             util.table_pipe(row:get(), 24, true),
-            util.table_border(row:increment(), 'delimiter', { 11, 11 }),
+            util.table_delimiter(row:increment(), { 11, 11 }),
             util.table_pipe(row:increment(), 0, false),
             util.table_pipe(row:get(), 12, false),
             util.table_pipe(row:get(), 24, false),
-            util.table_border(row:get(), 'below', { 11, 11 }),
+            util.table_border(row:get(), false, { 11, 11 }),
         })
 
         local actual = util.get_actual_marks()
@@ -96,7 +97,7 @@ describe('table.md', function()
             util.table_pipe(row:get(), 12, true),
             util.inline_code(row:get(), 14, 25),
             util.table_pipe(row:get(), 37, true),
-            util.table_border(row:increment(), 'delimiter', { 11, 24 }),
+            util.table_delimiter(row:increment(), { 11, { 23, 1 } }),
             util.table_pipe(row:increment(), 0, false),
             util.inline_code(row:get(), 2, 12),
             util.table_pipe(row:get(), 13, false),
@@ -106,15 +107,15 @@ describe('table.md', function()
 
         vim.list_extend(expected, {
             util.heading(row:increment(2), 1),
-            util.table_border(row:increment(2), 'above', { 11, 11 }),
+            util.table_border(row:increment(2), true, { 11, 11 }),
             util.table_pipe(row:get(), 0, true),
             util.table_pipe(row:get(), 12, true),
             util.table_pipe(row:get(), 24, true),
-            util.table_border(row:increment(), 'delimiter', { 11, 11 }),
+            util.table_delimiter(row:increment(), { 11, 11 }),
             util.table_pipe(row:increment(), 0, false),
             util.table_pipe(row:get(), 12, false),
             util.table_pipe(row:get(), 24, false),
-            util.table_border(row:get(), 'below', { 11, 11 }),
+            util.table_border(row:get(), false, { 11, 11 }),
         })
 
         local actual = util.get_actual_marks()
@@ -144,23 +145,23 @@ describe('table.md', function()
 
         vim.list_extend(expected, {
             util.heading(row:get(), 1),
-            util.table_border(row:increment(2), 'above', { 11, 24 }),
+            util.table_border(row:increment(2), true, { 11, 24 }),
             table_row(row:get(), 38, '│ Heading 1 │ `Heading 2`            │', true),
             util.inline_code(row:get(), 14, 25),
-            util.table_border(row:increment(), 'delimiter', { 11, 24 }),
+            util.table_delimiter(row:increment(), { 11, { 23, 1 } }),
             table_row(row:increment(), 40, '│ `Item 行` │ [link](https://行.com) │', false),
             util.inline_code(row:get(), 2, 12),
             util.link(row:get(), 15, 38, 'web'),
-            util.table_border(row:get(), 'below', { 11, 24 }),
+            util.table_border(row:get(), false, { 11, 24 }),
         })
 
         vim.list_extend(expected, {
             util.heading(row:increment(2), 1),
-            util.table_border(row:increment(2), 'above', { 11, 11 }),
+            util.table_border(row:increment(2), true, { 11, 11 }),
             table_row(row:get(), 25, '│ Heading 1 │ Heading 2 │', true),
-            util.table_border(row:increment(), 'delimiter', { 11, 11 }),
+            util.table_delimiter(row:increment(), { 11, 11 }),
             table_row(row:increment(), 25, '│ Item 1    │ Item 2    │', false),
-            util.table_border(row:get(), 'below', { 11, 11 }),
+            util.table_border(row:get(), false, { 11, 11 }),
         })
 
         local actual = util.get_actual_marks()