+++ CHANGELOG.md
@@ -12,11 +12,14 @@
   [965c222](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/965c222076b2d289ed498730845d533780f3c7c7)
 - code language name [#205](https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/205)
   [18c7ef7](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/18c7ef71fb4b8d83cb0160adc9127fc4d65ca42e)
+- anti conceal per component [#204](https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/204)
+  [fb6b3d1](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/fb6b3d145e5e12b838c0b84124354802f381b1af)
 
 ### Bug Fixes
 
 - bullet point right padding priority [#199](https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/199)
   [b02c8ef](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/b02c8ef72b10537a346556696b6e7fb354c8771f)
+- window offset bottom calculation [e13ac2c](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/e13ac2c05d2f081453db1451ec07fbd8be33ceec)
 
 ## 7.3.0 (2024-10-04)